pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-sre"
version = "0.1.0"
description = "SRE agent with full observability"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.30.0",
    "ddtrace>=2.20.1",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
